correct module naming
