+++ lib/sms_sender_smsgw/mobile_number_normalizer.rb
@@ -1,0 +1,10 @@
+module SmsSenderSmsgw
+  module MobileNumberNormalizer
+    def self.normalize_number(n)
+      while n.starts_with?('+') || n.starts_with?('0')
+        n.slice!(0)
+      end
+      return n
+    end
+  end
+end
